requirements.txt
Package            Version  
------------------ ---------
certifi            2021.10.8
charset-normalizer 2.0.7    
idna               3.3      
loguru             0.5.3    
pip                20.0.2   
pkg-resources      0.0.0    
python-dotenv      0.19.1   
requests           2.26.0   
setuptools         44.0.0   
urllib3            1.26.7   
WooCommerce        3.0.0    
